+++ flatlaf-core/src/main/java/com/formdev/flatlaf/ui/FlatSplitPaneUI.java
@@ -84,7 +84,7 @@ import com.formdev.flatlaf.util.UIScale;
  */
 public class FlatSplitPaneUI
 	extends BasicSplitPaneUI
-	implements StyleableUI
+	implements StyleableUI, FlatTitlePane.TitleBarCaptionHitTest
 {
 	@Styleable protected String arrowType;
 	/** @since 3.3 */ @Styleable protected Color draggingColor;
@@ -225,6 +225,15 @@ public class FlatSplitPaneUI
 		// paint divider style (e.g. grip)
 		if( divider instanceof FlatSplitPaneDivider )
 			((FlatSplitPaneDivider)divider).paintStyle( g, x, y, width, height );
+	}
+
+	//---- interface FlatTitlePane.TitleBarCaptionHitTest ----
+
+	/** @since 3.4 */
+	@Override
+	public Boolean isTitleBarCaptionAt( int x, int y ) {
+		// necessary because BasicSplitPaneDivider adds some mouse listeners for dragging divider
+		return null; // check children
 	}
 
 	//---- class FlatSplitPaneDivider -----------------------------------------